some changes

+++ src/main/java/com/scorp/socialmedia/repository/LikeRepository.java
@@ -3,5 +3,5 @@ package com.scorp.socialmedia.repository;
 import com.scorp.socialmedia.entity.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikeRepository extends JpaRepository<Like, Long> {
+public interface LikeRepository extends JpaRepository<Like, Integer> {
 }
